+++ src/vue/result/main.js
@@ -3,7 +3,8 @@ import App from './App'
 import ElementUI from 'element-ui';
 import locale from 'element-ui/lib/locale/lang/en'
 Vue.use(ElementUI, { locale });
-
+import Contextmenu from "./component/Contextmenu"
+Vue.use(Contextmenu);
 import UmyTable from 'umy-table'
 import 'umy-table/lib/theme-chalk/index.css';
 import '@/../public/theme/auto.css'